split axis in separate file
